Refactoring to remove things from the spring boot controller
